@@ -2,14 +2,17 @@ source 'https://rubygems.org'
 
 # Core
 gem 'pg'
-gem 'rails', '4.1.9'
+gem 'rails', '~> 4.2.9'
+gem 'responders', '~> 2.2.0'
+
 
 # Authentication
-gem 'devise', '~> 3.4.1'
+gem 'devise', '~> 3.5.10'
 gem 'devise_ldap_authenticatable', '~> 0.8', github: 'yeti-switch/devise_ldap_authenticatable'
 gem 'activeldap'
 gem 'net-ldap', '~> 0.3.1'
 gem 'd3-rails'
+gem 'knock', '~> 2.1.1'
 
 # ActiveAdmin
 gem 'ransack', '~> 1.4.0'
@@ -75,12 +78,12 @@ end
 group :development, :test do
   gem 'thin'
 
-  gem 'rspec-rails'
+  gem 'rspec-rails', '~> 3.4.2'
   gem 'factory_girl_rails'
   gem 'database_cleaner'
   gem 'rspec_api_documentation', '~> 5.0.0'
-  gem 'apitome', '~> 0.1.0'
 end
 gem 'unicorn-worker-killer'
 gem 'ipaddress'
+gem 'apitome', '~> 0.1.0'
 
