@@ -1,7 +1,13 @@
 require 'spec_helper'
 
 describe Api::Rest::Private::DialpeersController, type: :controller do
-  before { request.accept = 'application/json' }
+  let(:user) { create :admin_user }
+  let(:auth_token) { ::Knock::AuthToken.new(payload: { sub: user.id }).token }
+
+  before do
+    request.accept = 'application/json'
+    request.headers['Authorization'] = auth_token
+  end
 
   describe 'GET index' do
     let!(:dialpeers) { create_list :dialpeer, 2 }
